mdm-cdc-producer/pyproject.toml
[project]
name = "mdm-cdc-producer"
version = "0.1.0"
description = "Consumes Debezium CDC events and publishes MDM topics"
requires-python = ">=3.11"
dependencies = [
    "kafka-python==2.0.2",
]

[project.scripts]
mdm-cdc-producer = "app.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
